Remove redundant entries from harmony_to_diagram.xml
This data is read into an `unordered_map` which does not allow duplicates. Only the first entries should have been used. However, I found some discrepancies between macOS and Linux. For an Ab chord, macOS was using the first entry as expected. Linux was using the second.
I have a feeling this is due to a faulty `hash` implementation for `String`. It seemed like on Linux the collision wasn't noticed in `insert` and the original entry overwritten. I couldn't get this working and removing redundant entries seemed a sensible thing to do anyway.